+++ src/main/java/com/mohaemukzip/mohaemukzip_be/domain/recipe/controller/RecipeController.java
@@ -30,15 +30,6 @@ public class RecipeController {
     private final RecipeCommandService recipeCommandService;
     private final RecipeSearchService recipeSearchService;
 
-    @PostMapping
-    @Operation(summary = "레시피 저장 API", description = "특정 video_id를 가진 유튜브 영상에 관한 레시피를 저장합니다.")
-    public ApiResponse<RecipeResponseDTO.RecipeCreateResponse> createRecipe(
-            @RequestBody RecipeRequestDTO.RecipeCreateRequest request
-    ) {
-        Long recipeId = recipeCommandService.saveRecipeByVideoId(request.getVideoId());
-        return ApiResponse.onSuccess(new RecipeResponseDTO.RecipeCreateResponse(recipeId));
-    }
-
 
     @GetMapping("/{recipeId}")
     @Operation(summary = "세부 레시피 조회 API", description = "특정 videoId에 속하는 레시피에 관한 정보를 조회합니다.")
@@ -51,17 +42,7 @@ public class RecipeController {
         );
     }
 
-    @PostMapping("{recipeId}/summary")
-    @Operation(summary = "요약 레시피 생성 API", description = "특정 RecipeId에 속하는 레시피의 조리법을 요약해서 저장합니다.")
-    public ApiResponse<RecipeResponseDTO.SummaryCreateResponse> createSummary(
-            @PathVariable Long recipeId
-    ) {
-        var result = recipeCommandService.createSummary(recipeId);
-        return ApiResponse.onSuccess(
-                new RecipeResponseDTO.SummaryCreateResponse(result.summaryExists(), result.stepCount())
-        );
-    }
-    
+
     @PostMapping("/{recipeId}/complete")
     @Operation(summary = "요리 완료 API", description = "특정 RecipeId에 속하는 레시피를 CookingRecord에 저장하고, 해당 레시피의 난이도를 갱신합니다.")
     public ApiResponse<RecipeResponseDTO.CookingRecordCreateResponseDTO> createCookingRecord(
